Add requirements.txt
Using a `requirements.txt` file is a standard way to track and install
python dependencies. This allows one to keep track of all the
dependencies and versions as well as easily and quickly install them
when setting up a new instance or in a CI system.

I also changed the instruction to create a `.venv` directory within the
project directory for the virtual environment. This is also fairly
standard practice and helps to keep things organized if you have
multiple projects. It is ignored in `.gitignore`.

Finally, I added a step to create a `tracebase` postgres user, which is
not really necessary for development sites, but will be more similar to
how things are done in production.