Added Protocol and MSRun models.
Also created a stub placeholder model for Sample, since MSRun linked to it.
I made and applied migrations, which incidentall, included a change to hmdb_id from the Compound model, because Lance had modified it to not allow it to be blank.

Files checked in: DataRepo/models.py DataRepo/migrations/0002_auto_20210322_1629.py