Ensure eol for all files is lf
Specifying an `.editorconfig` will ensure the Super Linter checks that
files conform to the rules specified. This is done via
[editorconfig-checker](https://github.com/editorconfig-checker/editorconfig-checker/)

The config specifies that all files end with a line-feed (unix style
text files) and that there is a final new line. Also the charset should
be `utf-8` and there should be no trailing whitespace.  One exception is
that `.tsv` files may have trailing whitespace to indicate empty
columns.