Add load_samples command (#34)
* Add load_samples command

Create a `SampleTableLoader` class in `utils.py`

The `load_samples` command takes the sample table and a header
definition (in yaml) and loads the data into the database.

Any tissues, studies, animals, or samples that already exist in the
database will be skipped. If they do not exist, they will be created.

Compounds must be loaded first, and all tracer compounds must exist in
the database prior to loading.

The header definition allows some fields to be left out (indicated by an
empty value in the yaml file). Some fields are required by the model,
and thus must not be blank. The model definition was updated so that
fields that were allowed to be null are also allowed to be blank so that
the Django validation can be used to check the data prior to saving.

New required packages were added to `requirements/common.txt`.

* Fix typo in comment in DataRepo/models.py

Co-authored-by: John C. Matese <jcmatese@princeton.edu>

* Fix formatting and namedtuple definition

* Ensure eol for all files is lf

Specifying an `.editorconfig` will ensure the Super Linter checks that
files conform to the rules specified. This is done via
[editorconfig-checker](https://github.com/editorconfig-checker/editorconfig-checker/)

The config specifies that all files end with a line-feed (unix style
text files) and that there is a final new line. Also the charset should
be `utf-8` and there should be no trailing whitespace.  One exception is
that `.tsv` files may have trailing whitespace to indicate empty
columns.

* Use dashes in named arguments for commands

Using dashes for named arguments provides a simpler interface for users
and is preferred by the GNU Style Guide.

* Fix argument name in load_samples test

* Change ATOM to ELEMENT

Element is the correct term, since we mean to indicate Hydrogen,
Carbon, Oxygen, etc.

* Interpret empty string for tissue as blank sample

Sample records will be skipped if the Tissue field is an empty string.
This seemed an intuitive and consistent way to indicate a "BLANK"
sample, which for now we are not loading into the database.

Co-authored-by: John C. Matese <jcmatese@princeton.edu>