Preserved the form when invalid.
Before this change, if the form was invalid, the whole thing was wiped out - all selections and entered values.  Now, it is restored, however it does not yet show a validation error.  I was able to prevent the validation error from causing an internal server error by removing 'label' and 'div' from the cloned element loop that was used to restore form values.  The label I
'd added to show the error was getting into the loop and something was complaining that an object was not defined.  I'd tried adding a name to the element, which is what seemed to cause the error, but that didn't fix it.  Rather than work that out, I just removed it, because it wasn't necessary to restore the form values.

Files checked in: DataRepo/forms.py DataRepo/templates/DataRepo/peakgroups_results3.html DataRepo/views.py static/js/nested_forms.js