The forms are now submitted and converted into an object.
I think that the searches can be executed, but I have not tested that yet.
Unfortunately, there is somewhat of a kludge in this code.  It uses an arbitrary output format key to send all the form data and the form type is obtained from the pos field.  So the wrong is_valid is likely being called for 1 of the 2 forms.  I should only check validity on the selected format and ignore the unselected one(s).

Files checked in: DataRepo/forms.py DataRepo/multiforms.py DataRepo/templates/DataRepo/search_advanced.html DataRepo/views.py static/js/hierarchical_formsets.js