Successfully extracted the prefix/formname variables.
I had tried setting a prefix (from the multiforms.py file and tracking it in the formname key in the qry data structure.  It worked for the initial search page, but on the results page, all the forms had the generic "form" prefix in their names/IDs and I could not figure out how to set that the way it did in the initial load, so I just reverted to using the generic "form" prefix and everything worked.  I had intended to use that prefic to somehow create a single management form, but that didn't work the way I expected, so it was pointless anyway.

Files checked in: DataRepo/multiforms.py DataRepo/tests/test_views.py DataRepo/views.py static/js/hierarchical_formsets.js