Download search & search_basic redirect to search_advanced.
A download button now appears at the top of the advanced search results.  Clicking it downloads the results currently displayed on the page.

It needs to be tested to work with the first page of browsing data.

I also caused search_basic to go to the advanced search page.  To make this work seamlessly, I had to add the various table IDs to the select lists of the advanced search so that the search form would be correctly populated.  It works by constructing a qry object from the single query.

A bunch of the code is currently duplicated and needs to be streamlined.

Files checked in: DataRepo/forms.py DataRepo/templates/DataRepo/search_advanced.html DataRepo/templatetags/advsrch_tags.py DataRepo/urls.py DataRepo/views.py DataRepo/templates/DataRepo/search_advanced.tsv