Accept xlsx files as animal and sample tables
The template animal and sample tables are in a spreadsheet and
researchers find it simpler to provide spreadsheet files instead of csv
files for the animal and sample data, thus accepting those directly is
much simpler.

The default table headers were updated to reflect the standardized
naming conventions and match those in the animal an sample template
files provided to the researchers.