Resolves issue #135 - form field text entry delay.
The real fix/workaround was the addition of the placeholder input attribute, which prevents Safari Autofill from causing a delay after clicking in the field and also results in suggesting phone numbers to enter into the field from their contact card on macOS.  Note, other browsers I tested did not have these issues.  Many suggested solutions on google/stack were tried and this was the only one that worked.

I also reduced the number of calls to getElementById in the javascript, thinking initially that it was that that caused the lag.  It did not end up being the cause of the lag, but is more efficient none-the-less, so I kept it.

Files checked in: DataRepo/forms.py static/js/hierarchical_formsets.js