Re-added the check on the real database for sample uniqueness and existence of researcher.
The test database appears to be populated. I was trying to figure out how to initialize it as empty, but could not figure it out.  However, I did confirm that the tests are not leaving the database modified.

Files checked in: DataRepo/example_data/data_submission_animal_sample_table.xlsx DataRepo/utils.py DataRepo/views.py
[skip ci]