When handoffs are performed for fields in tables that are related with the previous table in the key path in a many-to-many fashion, there is a check to ensute the correct value is retrieved to get the same results as you would get with the searched field that is not displayed.  However there was an assumption that records retrieved would only match the search term.  That assumption was incorrect, thus the records must be traversed to find the correct handoff field.
There may exist another latent bug that we should not run into in the currect model.  If there is more than 1 M:M relationship in the key path (which there currently is not), you could end up looking at either no records or an assorment of records not matching the search term depending on the query's or'ed structure.  It could even be possible currently given the right search structure, thouugh I think it's very unlikely.

Files checked in: DataRepo/views.py
[skip ci]