Added a counter to affected search results templates to count the effective number of rows in the results.
This had to be accomplished via javascript because unless you loop through all records twice, there was no way to put an accurate count at the top of the results.

It is interesting that if the search term is on a field in a many-to-many related table, the count is accurate, but if the search terms do not include terms from the many-to-many table, the record count  does not include extra records from the M:M table.  For example, if I search the PeakGroups format for tissue=brain, the record count is 28, but the number of rows in the table is actually 35, because a number of records are of animals that belong to 2 studies.

Files checked in: DataRepo/templates/DataRepo/search/results/peakdata.html DataRepo/templates/DataRepo/search/results/peakgroups.html DataRepo/templatetags/customtags.py
[skip ci]