Fix Animal.age and Animal.tracer validation
Animal.age is a DurationField and thus the minimum value must be a
`timedelta` and not an `int`.

Animal.tracer was allowed to be null, but not blank. Thus the database
would allow null values, but validation failed.

Added a test to create and validate an animal model that failed before this fix.

Fixes #259 and addresses parts of #180