Catch errors when loading compounds
Catch errors when loading compounds and raise an exception
This makes is easier for other commands (such as load_study) to catch
the issues and abort any ongoing transactions.

Additionally, move much of the output from the validation and data
loading functions to the command line management command to give more
control to the calling functions and allow for other usage.

Partially addresses #289