Make tracer_compound a required field
Ensure that Animals have a valid tracer_compound and
tracer_labeled_count value. This prevents other issues downstream.

Update Animal tests to account for this change

Update DataFrame tests to account for this change

Migration sets any missing values to `compound_id=5` which is `arginine`
most likely. It also sets the `tracer_labeled_count=6`. These were the
values for the study where the missing values were discovered, but
reloading of the data would ensure it is correct.

Resolves #312