Added mechanisms to track whether any cached values exist beneath a root record.
This greatly speeds up the loading script and loads during tests (even though they use DummyCache).  The basic idea is the use the first cached function of the root model as a proxy to infer whether any cached values exist.  Thus, whenever a cached value is set, the root record's first cached function is also set.  Later, when loading scripts are run, that proxy is checked.  If it's not cached, then the delete_cache cascade for that hierarchy is not traversed.

Files checked in: DataRepo/hier_cached_model.py DataRepo/utils.py