Simplification.
The code I used originally was similar to this, but while figuring out bugs in the code, I'd made it unnecessarily complicated before I figured out what the problem was.  I committed what I knew was working previously and then reverted the code to this simpler version, which also works - and is an order of magnitude faster.

Files checked in: DataRepo/views.py
[skip ci]