Used .annotate() to simulate a proper full join.
It works!  Currently, I have MeasuredCompound set up to split rows, but I only did it in the html template, not the tsv download.  Note that this basically returns duplicate root table records and always returns every matching M:M table record, so the template has to know which one belongs to which row by checking the annotated value against the value encountered in the nested loop in the template.  If it's a match, it prints.  Otherwise, it moves on.

Files checked in: DataRepo/compositeviews.py DataRepo/templates/DataRepo/search/results/peakgroups.html DataRepo/views.py
[skip ci]