Record counts now reflect the results of a join through a many-to-many table.
And, I fixed an issue I hadn't noticed before, where a query that results in multiple M:M related table records was rendered as duplicate rows in the html table.

I did that by adding an annotation of the linked table record's primary key value.

The end result is that if you set full_join to True for an M:M related table, only one representation of that table's values will be displayed on a row.

I created a template tag to retrieve the record or records for every row, of a M:M related table.  You don't even need a nested for loop, but I made the template tag always return a list so that the template code is the same whether it's for a full join or not.

Files checked in: DataRepo/compositeviews.py DataRepo/templates/DataRepo/search/results/peakgroups.html DataRepo/templatetags/customtags.py DataRepo/tests/test_compositeviews.py