Fixed another (hopefully the last) bug that existed at the intersection of MaintainedModel and the validation database.
The validation code had not operated on classes that inherit from MaintainedModel.  In the tests, it was loading a sample table with a lysine tracer into the default database.  Then it was loading a different  sample table in the validation code which had a single glucose tracer.  The validation code loads twice: once in debug mode (which raises an exception), and if not failed, once without debug mode (so that the data is there for the accucor validation.  Well, the validation load was performing mass auto-updates on a buffer that wasn't cleared out.  It was calling save on the old lysine object and on the glucose object.  One of them was creating an empty name, which went silently undetected until you try to run 2 such tests in 1 test run where it hits the unique contraint for trying to create a second empty name record.

Part of the problem is that when an exception occurs, the buffer was not cleared and the next load/validation attempt gets the old buffer.

I added calls to clear_update_buffer() in sample_table_loader for 2 exception cases that gets the tests to pass, but really what should be done is the sample table loader should wrap everything in a try and in a finally, clear the buffer and re-enable auto-updates.  I'll leave that to another issue.

Files checked in: DataRepo/models/infusate.py DataRepo/models/infusate_tracer.py DataRepo/models/maintained_model.py DataRepo/models/tracer.py DataRepo/models/tracer_label.py DataRepo/tests/test_views.py DataRepo/utils/sample_table_loader.py DataRepo/views/loading/validation.py
[skip ci]