Partially working.
The first click of the show last/previous checkboxes works in the show last case.  It is not working in the show previous case because it seems to set the search to is_last is true...

The row hiding needs to be disabled too.

Files checked in: DataRepo/formats/fluxcirc_dataformat.py DataRepo/templates/DataRepo/search/results/fcirc.html static/js/hierarchical_formsets.js
[skip ci]