A number of fixes for tests.
Edited the comment in DataRepo/views/loading/validation.py to better explain why the sample table spreadsheet is tested in debug mode first.
Changed the usage of get_or_create on Protocol to a get, then later call a create.  I strongly suspect that the database being operated on in the create in the get_or_create case was the test tracebase database even though the database in the call was the validation database.  I was able to insert prints in the django code from analyzing the trace that showed the database that the get works on is the test  validation database, but the create was complaining that a protocol record existed and that the attempt to create key 1 violated the unique constraint.  Since the validation database starts out empty, this made no sense.  The name in the existing record was an MSRun protocol, which should only exist in the test tracebase database, so I expect that django is switching databases between the get and the create.  My fix works around this by doing the operations separately.  The real solution will be in #569 where the router is overridden.
I changed the protocol loader to mimic the other load scripts WRT validation.  So it no longer loads into BOTH databases.  Thus, the validation now copies protocols in the tracebase database to the validation database anytime any files are validated.
I also changed the base class of ViewTests to TracebaseTransactionTestCase because after fixing things, I got known transaction management errors having to do with implicit wrapping of each test in an atomic transaction.  I've discussed this necessity before when I created that base class.
Once the transaction test case was used and the database switching was addressed, the existing tests for validating files failed.  They no longer contained the researcher errors they expected.  And I looked it over and I realized that there SHOULD BE NO researcher errors.  There never should have been.  I suspect that it all has to do with things operating on the default database because we haven't implemented a router.

As long as I can work out any remaining issues, even if things ultimately work, I may be amenable to turning off the validation view until the router is implemented.  There could be more unexpected issues.

Files checked in: DataRepo/tests/test_views.py DataRepo/utils/protocols_loader.py DataRepo/views/loading/validation.py