Added a way to discern between a blank form being created versus a populated form.
The initial units select list default value changes based on the initial default fld value, so it must be initialized after the form field loop completes, but only when we're loading a blank form.  Otherwise, the form field value for the units select list must be based on what's in the incoming query object, which is stored in the clone of the field form that was created in the loop.  before, the setting of the default value was put in a conditional that checks a different form field element for a different reason.  So I created a boolean and infer a blank form state based on the val field element.

Another thing that was addressed was that the notion of the default "native" field had previously been changed to "identity", but that update wasn't implemented everywhere it should have.  In fact, there should be a check to ensure every type has an "identity" units format.

Files checked in: DataRepo/formats/dataformat.py DataRepo/formats/fluxcirc_dataformat.py DataRepo/formats/peakdata_dataformat.py DataRepo/formats/peakgroups_dataformat.py DataRepo/tests/test_formats.py static/js/hierarchical_formsets.js