Test fix.
"native" was changed to "identity", which changes the return of getAllFieldUnitsChoices, which makes the value unique (name is arbitrary - only the value is important).  Just had to remove the second "identity" value entry in the tuple.  The only reason for that tested function is to be able to allow django to validate each form despite different value population.

Files checked in: DataRepo/tests/test_formats.py