Added AggregatedErrors to the load_study script.
Since there are multiple AggregatedErrors objects involved, I decided to handle things a bit differently, but similar in concept to how I was doing this already in the validation view.  I created 2 extra classes:

AggregatedErrorsSet - This is what handles raising an exception when either doing an actual load or Dry Run load.  It only reports the exceptions and raises the AggregatedErrrosSet exception.
MultiLoadStatus - This is an exception class, but it's only an exception because call_command doesn't provide a way to return anything.  It keeps the status of everything, whether it fails or not.

I also implemented part of #648.  It creates separate exceptions that includes consolidated exceptions from multiple files.  Currently, it includes missing compounds and missing samples.  I didn't do anything with missing tissues, because the tissues loader does not yet have AggregatedErrors incorporated.

Files checked in: DataRepo/management/commands/load_study.py DataRepo/utils/exceptions.py