Fixed an issue with a previous edit.
In a previous commit, I decided to raise an error when an AggregatedErrors object (or other exception) was set using the same load key.  I neglected to anticipate that this is expected to happen if there are errors in both the protocol load and the animal and sample table load, sinc ethe load key in both cases is the same file name.  So I:

- Added a method named `merge_object` to the AggregatedErrors class
- Added a test for the new method

Files checked in: DataRepo/tests/test_exceptions.py DataRepo/utils/exceptions.py