load_study should always run child load scripts in validate mode in order to catch as many errors as possible.
I noted that accucor loads all failed when there was an error from the sample table loader and I realized that that was because the validate flag was false.  If it had been true in the call to the sample table loader, the sample records would have not been rolled back and the related accucor loads would not have failed, yet once load_study finished, it would have rolled everything back due to the error in the sample table loader.

Files checked in: DataRepo/management/commands/load_study.py DataRepo/management/commands/load_study_set.py
[skip ci]