Fixed a bug in clear_update_buffer where the generation argument was being compared to a number when it was None.
There was another bug however, and that was setting the argument to the clear_update_buffer function that was a global list value.  The argument name was a global variable itself and it was evaluated before the global variable used to set it, so this argument value was never working and the function was never using the global filters if they were set.  We were using the global filters in the sample table loader to filter for label "name".  Since that was the only maintained series of fields in the tables loaded by the sample table loader, we did not experience any ill effects.

Files checked in: DataRepo/models/maintained_model.py DataRepo/models/maintained_model.py