Fixed a few bugs.
- I was incorrectly accessing the coordinator in a few places.
- _get_current_coordinator was doing something wrong - before I understood what I was doing.  It was calling the coordinator like a method, which I think was based on an article I was reading, but I ended up abandonning that article.
- I decided to not bother registering the class from the relation decorator.
- Looks like I crossed wires on the get_updater_dicts_by_model_name method.  It had the code for get_my_update_fields.

I was having an exception occur related to the new way of registering the classes, which is what I was using to perform dynamic queries.  It kept giving my a TypeError about a dict not being hashable.  I believe it had to do with the way I was saving the classes, but when I put a debug print in django's core code, it started working!?  Perhaps there was a stale python cache somewhere...

Files checked in: DataRepo/models/maintained_model.py