Made some label filter updates.
There was an error in the buffer size calculation and the label filter settings in many of the methods were defaulting to the coordinator defaults and I realized that (while it currently works fine because we don't use any custom filters) doing that was wrong in a number of the methods because of the passing of buffer contents from child to parent coordinators (which can have different filters).  If the current default filters are used in a parent, it could skip/miss buffered items that came from the child, so I set the defaults to no filters, unless filters were supplied to the method.

I also noted that if no filters are present, I don't want to use any potentially supplied filter_in value - I want to use True.

I tweaked the filter check in the coordinator init method to account for an empty list in addition to None.  This was an oversight when I changed the default due to None being problematic.

I also tweaked some of the comments and deleted some more vestigial deleted method comments.

Files checked in: DataRepo/models/maintained_model.py