Added a disable of caching updates to the LCMethod record migration.
This makes the migration MUCH faster.  I discovered the slowness because after my initial migration, I loaded new data on a different branch that was causing me errors, so I had to copy the method into the shell and run it manually.  It was taking forever and there were a lot of caching update messages, so I stopped it and added the caching update disable call and it finished in short order.

Files checked in: DataRepo/migrations/0016_protocols_to_lcmethods.py