Fixed a bug that occurs when an exception happens during the setup call to the ViewTests class from ViewNullTolerance tests that cascades to other test classes.
The original exception was expected - because I haven't updated the tests yet, but ViewNullTolerance tests was doing some under-the-hood stuff with coordinators in order to re-run all of the ViewTests without any of the maintained fields having values.  It calls a private method to artificially add a disabled coordinator, because a context manager would not have worked (I don't think).  Anyway, the expected exception was causing tearDownClass to not be called, so that coordinator wasn't getting popped off the stack.  So I just added an except block to remove the coordinator in the event of any exception.  Now we're back to only the expected errors.

Files checked in: DataRepo/tests/test_views.py
[skip ci]