Reverted unused changes to tracer_label.py and improved the unique constraints and field to header mappings.
I realized that the only purpose of the field to headermappings was to be able to reference file data when the field referenced in an error directly maps to a column.  In fact, this could probably be modified to handle multiple columns mapped to a single database model field.

I also realized that the unique constraint needed to have ID_KEY added to it, because all of the compound and isotope data could not be unique without it, if a tracer included the same carbon label details, but a different nitrogran label configuration.  Then I also added the same unique constraint, but I included the tracer name instead of the number.

Files checked in: DataRepo/loaders/tracers_loader.py DataRepo/models/tracer_label.py