Improved exception inheritance.
A lot of the exception code was repetitive.  All of the exceptions that were using `generate_file_location_string` and `summarize_int_list` were modified to inherit from a new base exception class named `InfileError`.

A feature was added to the base class to be able to manually specify where location information should be incorporated into a supplied message.  This is achieved with an "order" argument that corresponds to what information should be placed where formatting "" characters are in the supplied message.

I also noted that a number of the exception tests were put in the wrong test class, so I rearranged them to be in the appropriate classes.

I also created a new exception class named CompoundDoesNotExist in order to provide file context to the exception.

Files checked in: DataRepo/loaders/accucor_data_loader.py DataRepo/tests/utils/test_exceptions.py DataRepo/utils/exceptions.py