This was an attempt to download the file and render the results page in 1 go using a multipart stream.
The python component took about 2 hours, though I never confirmed the integrity of the file.  The rest of the day, I worked on getting the javascript on the page to submit the form and process the multipart stream (render a new page in a new tab and kick off the download of the excel file), and I got close, but I never got it to fully work.  I'll check in the template next.

This also includes a new requirement for the xlsxwriter package.

Files checked in: DataRepo/views/upload/validation.py requirements/common.txt
[skip ci]