Added a parse_isotope_label method to the (poorly named) infusate_name_parser.py utils file.
Some of the parsing exception names were changed to make these changes durable (so that the accucor loader still works).

Note, the PR associated with this change does not (by itself) need to merge into the converted table loader branch, but the changes in this PR and the next PR are all needed (along with the PRs thus far) by the peak annotations loader PR.  That's why this and the next PR are a part of the chain - because you can't merge into multiple PRs.