Cherry-picked necessary updates to TableLoader to support the coming StudyLoader class.
StudyLoader does not directly update any models.  It just calls other loaders, so it needed to be able to have an empty Models class attribute.  StudyLoader also needs to be able to raise 2 new exceptions: AggregatedErrorsSet and MultiLoadStatus.  The set exception basically organizes all of the AggregatedErrors sets from all the different loaders into a dict keyed on study sheet name.  So the wrapper for the derived loader classes' `load_data` methods needed to not grab those exceptions and push them onto an AggregatedErrors exception's list of exceptions.  It just needs to raise them when it sees the,

I also added a bunch of is_error and is_fatal arguments to the methods that handle exceptions, though I didn't end up actually using them.  The purpose at the time was to resolve an issue where animal loader warnings were somehow resulting in a rollback when it shouldn't have rolled back.  that problem however turned out to be that I wasn't supplying defer_rollback=True, so I didn't actually need to pass those new arguments.  But, it didn't hurt anything and it left the option open to treat exceptions how you want, so I left those changes in.