Fixed an issue that caused an exception during the migration check.
It turns out that for some reason, the migration check imports urls.py, and thus imports any views it references, and thereby any classes it imports... and if any of those classes have a database query to set a class attribute, it will execute BEFORE the database has been set up.  THAT causes a table to be reported as undefined.

I had tried to set static_choices for the view based on what currently was in the database for the AnimalLoader, so it encountered an exception during the migration check.  I hadn't run into it locally because I'd done the migration before making that change to static_choices.

So I added a new argument to the TableColumn constructor: current_choices (bool).  I made the static_choices into a private instance variable that doesn't evaluate during the instance creation, and created a static_choices property that should only ever evaluate when the static choices are needed.

Files checked in: DataRepo/loaders/animals_loader.py DataRepo/loaders/base/table_column.py