Improved the column metadata for both the animals and samples loaders.
This takes advantage of the new ability for header comments to automatically include a note when the column's drop-downs are populated by a column in another sheet (because I'd noted that I was repeatedly adding that comment as a "guidance" argument.

I also fleshed out the metadata based on what I had added in the `DataValidationView class before those loaders had been implkemented.