Added the ability for StudyLoader to convert study docs from the previous format to the current format.
Details:

- StudyLoader now inherits from ConvertedTableLoader instead of TableLoader.
- Added derived classes for each supported version of the study doc (versions 2 and 3).

These changes take advantage of functionality that tableLoader supplies, but overloads the concept of columns to apply to sheets.  It also overrides a few methods of ConvertedTableLoader, because ConvertedtableLoader was written with the assumption that it would always be converting to a single dataframe, which is what TableLoader assumes is its input as well.  Both of those classes need to be refactored to support the fact that the dataframe imbibed could alternatively be a dict of dataframes.

Files checked in: DataRepo/loaders/study_loader.py