Made dramatic efficiency improvements to the autofill-only mode of the build-a-submission page (when numerous compounds don't exist in the database).
It turns out that there were a number of aspects of the code that, when combined changes are implemented, there is a dramatic improvement to data-specific lag.

When the peak annotations file is large (many samples and many compounds), the unicorr format has the drawback that it hits pandas' iterrows pretty hard.  The time that pandas spends instantiating pd.Series objects becomes rather large.

Combine that with the fact that Compound.compound_matching_name_or_synonym has a worst-case running time when the compound is not found, so when the peak annotations file has numerous not-before-loaded compounds, that method also hits the overall running time pretty hard.

Lastly, you get a not insignificant running time hit when the PeakAnnotationsLoader buffers all those exceptions.  For the purposes of the autofill-only mode, we don't need those exceptions.  We only need them in validation mode, so I did 2 things to address this aspect:

1. Skip compounds we've seen before (so not failed lookup or exception occurs.
2. Added an argument to the methods involved to turn off error buffering.

It's also notable that, after the autofill step, the user will likely fill in the missing information in the Compounds sheet, so even if we buffer errors in the validation step, there shouldn't be (as many) missing compounds, so the problem should mostly go away at validation anyway.  Besides, in the validation step, we're only going to validate the study doc, so we won't be hitting these repeated exceptions anyway in the PeakAnnotationsLoader, so even if they don't fill the compounds sheet in, we won't hit this lag in the validation step.

I also added a debug print that prints the compounds added to the compounds sheet.

Files checked in: DataRepo/loaders/peak_annotations_loader.py DataRepo/tests/loaders/test_peak_annotations_loader.py DataRepo/views/upload/validation.py