I had to use a factory to create the BuildSubmissionForm class to avoid exceptions during the migrations check.
The problem is that the migrations process evaluates the class-level code before the database is built, so the db queries at the class level result in exceptions complaining about the models not existing.  This stack post put me on to the problem:

https://stackoverflow.com/a/56878154

I actually tried an alternate solution first.  I tried updating or setting the widget classes in the constructor.  While that functionally worked, something about that strategy broke the CSS on the page.  I couldn't figure out how to fix that, so I switched to a factory.

Files checked in: DataRepo/forms.py DataRepo/views/upload/submission.py