Added the ability to clone the form.
Note, I have not yet un-hidden the file input or changed any IDs or any other manipulations that will be necessary.  This just clones the form as-is and appends it as a child to the submission form.  It DOES copy the values entered into the sequence metadata to the right of the drop area.

Files checked in: DataRepo/templates/submission/submission.html static/js/file_list_drop_area.js static/js/submission_start.js
[skip ci]