Initial working version of drag-and-drop peak annotation file functionality that clones forms using a template empty form.
Note, this breaks the validate page and there are a number of things that DO NOT WORK as well as the ideal.  For example, submitting (valid or not) does not re-populate the forms (even though it *could* repopulate the template form) and errors do not appeat associated with any added forms or fields.

In order to avoid Safari autofill issues, this was written to provide a regular form to the template and that regular form is cloned in javascript.  Upon submit, the accumulated forms are manipulated to mimmic a django FormSet.  This should all be re-done, but is currently functional at an essential functionality level.

Details:

- Javascript was added to static/js/submission_start.js to manipulate forms into a formset.
- Cleaned up debug prints and vestigial code.

Files checked in: DataRepo/forms.py DataRepo/models/msrun_sequence.py DataRepo/views/upload/submission.py DataRepo/widgets.py static/css/drop_area.css static/js/file_list_drop_area.js static/js/submission_start.js