Removed the Sequence Name column and added a Common Samples column (along with 2 other metadata columns for sample count and a short set of example samples).
This supports only 1 peak group compound representation per sample as opposed to 1 per sample and sequence.  I could have left sequence and made it a delimited column, but decided that I could add it back later.  Right now, the priority is the basic functionality.

I changed the returned dict to be keyed on individual sample name instead of the combo and instead of the sequence.  It will make the lookup easier.

I intend to hide the common samples column, because it will be a mess to look at for the user, but it's utility is for the unique constraint (and to support multiple sequences in a file with difference sets of samples).

Files checked in: DataRepo/data/tests/multiple_representations/mult_reps.tsv DataRepo/loaders/peak_group_conflicts.py DataRepo/tests/loaders/test_peak_group_conflicts.py